backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
autogen-agentchat>=0.5.7
autogen-ext[openai]>=0.5.7
pydantic>=2.10.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.1.0
aiofiles>=23.2.1
